gate-watcher/src/pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "adafruit-ssd1306>=1.6.2",
    "flask>=3.1.2",
    "gpiozero>=2.0.1",
    "pigpio>=1.78",
    "pillow>=11.3.0",
    "requests>=2.32.5",
    "rpi-gpio>=0.7.1",
    "sense-hat>=2.6.0",
    "smbus>=1.1.post2",
]
